+++ segtypes/n64/header.py
@@ -17,7 +17,7 @@ class N64SegHeader(N64Segment):
         
         return f".{typ} {dstr} /* {comment} */"
 
-    def get_out_path(self) -> Path:
+    def out_path(self) -> Path:
         return options.get_asm_path() / self.dir / f"{self.name}.s"
 
     def split(self, rom_bytes):
@@ -47,7 +47,7 @@ class N64SegHeader(N64Segment):
         header_lines.append(self.get_line("byte", rom_bytes[0x3F:0x40], "Version"))
         header_lines.append("")
 
-        src_path = self.get_out_path()
+        src_path = self.out_path()
         src_path.parent.mkdir(parents=True, exist_ok=True)
         with open(src_path, "w", newline="\n") as f:
             f.write("\n".join(header_lines))
